+++ presentation/src/main/java/com/softpie/karabiner/component/button/KarabinerButton.kt
@@ -8,32 +8,22 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
-import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.softpie.karabiner.R
 import com.softpie.karabiner.component.modifier.karaOuterShadow
 import com.softpie.karabiner.component.theme.Headline
 import com.softpie.karabiner.component.theme.KarabinerColor
 import com.softpie.karabiner.component.theme.KarabinerRadius
 import com.softpie.karabiner.component.theme.KarabinerTheme
-import com.softpie.karabiner.component.theme.Label
 import com.softpie.karabiner.component.theme.contentColorFor
 import com.softpie.karabiner.component.theme.gradient
-
-sealed class ButtonType(val buttonColor: Color, val disableColor: Color = KarabinerColor.Gray200) {
-    object Black: ButtonType(buttonColor = KarabinerColor.Black)
-    object Gray: ButtonType(buttonColor = KarabinerColor.Gray100)
-    object White: ButtonType(buttonColor = KarabinerColor.White)
-    object Transparent: ButtonType(buttonColor = KarabinerColor.Transparent)
-}
 
 @Composable
 fun KarabinerButton(
@@ -46,7 +36,7 @@ fun KarabinerButton(
     shape: Shape = KarabinerTheme.shape.semiLarge,
     border: BorderStroke? = null,
     interactionSource: MutableInteractionSource = remember { MutableInteractionSource() },
-    contentPadding: PaddingValues = PaddingValues(vertical = 14.dp, horizontal = 24.dp),
+    contentPadding: PaddingValues = PaddingValues(vertical = 12.dp, horizontal = 24.dp),
     onClick: () -> Unit,
 ) {
     val colors = if (karabinerable)
@@ -91,9 +81,8 @@ fun KarabinerButton(
                 .then(modifier),
         ) {
             Column {
-                Icon(painter = painterResource(id = R.drawable.ic_list), contentDescription = null)
-                Label(
-                    modifier = Modifier,
+                Headline(
+                    modifier = Modifier.align(Alignment.CenterHorizontally),
                     text = text,
                     textColor = if (karabinerable) KarabinerColor.White else contentColorFor(
                         if (enabled) type.buttonColor
@@ -105,3 +94,26 @@ fun KarabinerButton(
     }
 }
 
+@Preview
+@Composable
+fun ButtonPreview() {
+
+    Column {
+
+        KarabinerButton(text = "신고하기") {
+
+        }
+        KarabinerButton(text = "신고하기", karabinerable = true) {
+
+        }
+        KarabinerButton(text = "신고하기", type = ButtonType.Gray) {
+
+        }
+        KarabinerButton(text = "신고하기", type = ButtonType.Gray, enabled = false) {
+
+        }
+        KarabinerButton(text = "신고하기", type = ButtonType.Transparent) {
+
+        }
+    }
+}
