+++ presentation/src/main/java/com/softpie/karabiner/ui/signup/name/screen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.unit.dp
 import androidx.core.view.WindowCompat
 import androidx.navigation.NavController
-import com.softpie.karabiner.component.button.KarabinerBottomButton
+import com.softpie.karabiner.component.button.KarabinerButton
 import com.softpie.karabiner.component.textfield.KarabinerTextField
 import com.softpie.karabiner.component.theme.Headline
 import com.softpie.karabiner.component.theme.KarabinerTheme
@@ -83,7 +83,7 @@ fun SignupNameScreen(
             )
         }
         Spacer(modifier = Modifier.weight(1f))
-        KarabinerBottomButton(
+        KarabinerButton(
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(horizontal = if (keyboardState) 0.dp else 24.dp),
